refactor: add components folder and update npm packages
